add authorisation by id for updationg and deliting user

+++ Backend/Web_API/Controllers/UserController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using BLL.Sevices;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.IdentityModel.JsonWebTokens;
 using Web_API.DTOs;
 
 namespace Web_API.Controllers
@@ -15,42 +16,76 @@ namespace Web_API.Controllers
     {
         // GET: api/<UsersController>
         [HttpGet]
-        public async Task<IEnumerable<UserInfoRequestDto>> GetUsers()
+        public async Task<IEnumerable<UserInfoDto>> GetUsers()
         {
             var users = await userService.GetAllUsersAsync();
-            var output = mapper.Map<IEnumerable<UserInfoRequestDto>>(users);
+            var output = mapper.Map<IEnumerable<UserInfoDto>>(users);
             return output;
         }
 
         // GET api/<UsersController>/5
         [HttpGet("{id}")]
-        public async Task<UserInfoRequestDto> GetUserById(int id)
+        public async Task<UserInfoDto> GetUserById(int id)
         {
             var user = await userService.GetUserByIdAsync(id);
-            var output = mapper.Map<UserInfoRequestDto>(user);
+            var output = mapper.Map<UserInfoDto>(user);
             return output;
         }
 
         [HttpGet("username/{userName}")]
-        public async Task<UserInfoRequestDto> GetUserByUserName(string userName)
+        public async Task<UserInfoDto> GetUserByUserName(string userName)
         {
             var user = await userService.GetUserByUserNameAsync(userName);
-            var output = mapper.Map<UserInfoRequestDto>(user);
+            var output = mapper.Map<UserInfoDto>(user);
             return output;
         }
 
         // PUT api/<UsersController>/5
         [HttpPut("{id}")]
-        public async Task UpdateUser(int id, [FromBody] UserReguestDto value)
+        public async Task<IActionResult> UpdateUser(int id, [FromBody] UserDto value)
         {
+            if (User?.Claims == null)
+            {
+                return Unauthorized();
+            }
+
+            var claim = User.Claims.FirstOrDefault(c => c.Type == "Id");
+            if (claim == null)
+            {
+                return Unauthorized();
+            }
+
+            var authorised = int.TryParse(claim.Value, out var userId);
+            if (!authorised || id != userId)
+            {
+                return Forbid();
+            }
             var userModel = mapper.Map<UserModel>(value);
             await userService.UpdateUserAsync(id, userModel);
+            return Ok(userModel);
         }
 
         // DELETE api/<UsersController>/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(int id)
         {
+            if (User?.Claims == null)
+            {
+                return Unauthorized();
+            }
+
+            var claim = User.Claims.FirstOrDefault(c => c.Type == "Id");
+            if (claim == null)
+            {
+                return Unauthorized();
+            }
+
+            var authorised = int.TryParse(claim.Value, out var userId);
+            if (!authorised || id != userId)
+            {
+                return Forbid();
+            }
+            
             await userService.DeleteUserAsync(id);
             return Ok();
         }
